* added remove function and optimized performance

+++ Fallen-8/Index/Spatial/IMBP.cs
@@ -23,9 +23,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-using System;
-using System.Linq;
-using System.Text;
 
 namespace Fallen8.API.Index.Spatial
 {
@@ -37,6 +34,6 @@ namespace Fallen8.API.Index.Spatial
         /// <summary>
         /// gets coordinates of point
         /// </summary>
-       Double[] Coordinates { get; }
+       float[] Coordinates { get; }
     }
 }
